+++ lib/app.ts
@@ -23,8 +23,6 @@ import {
 import { LocatorEngine, PathSlugStrategy } from "./locator/mod.ts";
 import { MdPageHandler } from "./content/mod.ts";
 import {
-  type ManagedPageDuplicator,
-  type ManagedPageRenamer,
   MemoryPageRepository,
   type PageDeliverer,
   PageHttpAdapter,
@@ -106,8 +104,6 @@ export interface AppServices {
   pages:
     & PageHttpApplication
     & PageDeliverer
-    & ManagedPageRenamer
-    & ManagedPageDuplicator
     & PublicPageViewer
     & PublicPageLister
     & PublicPageExplorer;